avoid broken += operator, bogus use of const

+++ src/test/run-pass/auto-ref-slice-plus-ref.rs
@@ -23,7 +23,7 @@ impl<'self> MyIter for &'self [int] {
 
 impl<'self> MyIter for &'self str {
     fn test_imm(&self) { assert_eq!(*self, "test") }
-    fn test_const(&const self) { assert_eq!(self[0], 't') }
+    fn test_const(&const self) { assert_eq!(self[0], 't' as u8) }
 }
 
 pub fn main() {
